requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
flask-swagger-ui==4.11.1
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==1.5.3
pymongo==3.10.1
requests==2.28.2
simplejson==3.18.4
six==1.14.0
urllib3==1.26.15
Werkzeug==1.0.1
xmltodict==0.13.0
